cli: support custom output dir
