+++ packages/cli/src/bin/pipcook-job.ts
@@ -1,11 +1,9 @@
 #!/usr/bin/env node
 
-import { ChildProcess } from 'child_process';
 import program from 'commander';
-import { get, listen } from '../request';
+import { get } from '../request';
+import { run } from '../pipeline';
 import { route } from '../router';
-import { tail } from '../utils';
-import { tunaMirrorURI } from '../config';
 import { ora } from '../utils';
 
 const PipelineStatus = [ 'creating', 'running', 'success', 'fail' ];
@@ -16,42 +14,6 @@ async function list(): Promise<void> {
     item.status = PipelineStatus[item.status];
     return item;
   }), [ 'id', 'status', 'evaluatePass', 'createdAt' ]);
-}
-
-async function run(id: string, opts: any): Promise<void> {
-  const spinner = ora();
-  const params = {
-    cwd: process.cwd(),
-    pipelineId: id,
-    pyIndex: opts.tuna ? tunaMirrorURI : undefined
-  };
-  if (!opts.verbose) {
-    const job = await get(`${route.job}/run`, params);
-    spinner.succeed(`create job(${job.id}) succeeded.`);
-  } else {
-    let stdout: ChildProcess, stderr: ChildProcess;
-    spinner.start(`start running ${id}...`);
-    await listen(`${route.job}/run`, params, {
-      'job created': (e: MessageEvent) => {
-        const job = JSON.parse(e.data);
-        spinner.succeed(`create job(${job.id}) succeeded.`);
-        stdout = tail(job.id, 'stdout');
-        stderr = tail(job.id, 'stderr');
-      },
-      'job finished': (e: MessageEvent) => {
-        const job = JSON.parse(e.data);
-        spinner.succeed(`job(${job.id}) is finished with ${e.data}`);
-        stdout?.kill();
-        stderr?.kill();
-      },
-      'error': (e: MessageEvent) => {
-        spinner.fail(`occurrs an error ${e.data}`);
-        stdout?.kill();
-        stderr?.kill();
-        process.exit(1);
-      }
-    });
-  }
 }
 
 async function remove(): Promise<void> {
@@ -82,6 +44,7 @@ program
   .command('run <pipeline>')
   .option('--verbose', 'prints verbose logs')
   .option('--tuna', 'use tuna mirror to install python packages')
+  .option('--output', 'the output directory name', 'output')
   .action(run)
   .description('run a job from a pipeline id');
